# [ARM CPU] Enable FP16 kernels for GQA op 