--- a/onnxruntime/contrib_ops/cpu/bert/attention_helper.h
+++ b/onnxruntime/contrib_ops/cpu/bert/attention_helper.h
@@ -17,93 +17,20 @@ namespace onnxruntime {
 namespace contrib {
 
 template <typename T>
-void ComputeSmoothSoftmaxInplace(T* score, int N, int D, ThreadPool* tp) {
-  ThreadPool::TryParallelFor(tp, N, D * 2.0, [&](std::ptrdiff_t begin, std::ptrdiff_t end) {
-    for (std::ptrdiff_t j = begin; j != end; ++j) {
-      float* x = reinterpret_cast<T*>(score) + j * D;
-      float* y = x;
-
-      float max = -std::numeric_limits<float>::infinity();
-      for (int i = 0; i < D; i++) {
-        if (max < x[i])
-          max = x[i];
-      }
-
-      if (max < 0.0f) {
-        max = 0.0f;
-      }
-
-      for (int i = 0; i < D; i++) {
-        y[i] = expf(x[i] - max);
-      }
-
-      double sum = 0.0;
-
-      for (int i = 0; i < D; i++) {
-        sum += x[i];
-      }
-
-      sum += exp(static_cast<double>(-max));
-
-      for (int i = 0; i < D; i++) {
-        y[i] = x[i] / (float)sum;
-      }
-    }
-  });
-}
-
-template <>
-inline void ComputeSmoothSoftmaxInplace(float* score, int N, int D, ThreadPool* tp) {
+inline void ComputeSmoothSoftmaxInplace(T* score, int N, int D, ThreadPool* tp) {
   MlasComputeSoftmax(score, score, N, D, false, true, tp);
 }
 
 template <typename T>
-void ComputeAttentionSoftmaxInplace(T* score, int N, int D, ThreadPool* tp) {
-  ThreadPool::TryParallelFor(tp, N, D * 2.0, [&](std::ptrdiff_t begin, std::ptrdiff_t end) {
-    for (std::ptrdiff_t j = begin; j != end; ++j) {
-      float* x = reinterpret_cast<T*>(score) + j * D;
-      float* y = x;
-
-      // e^x is represented as infinity if x is large enough, like 100.f.
-      // Infinity divided by Infinity is a NAN. Thus, softmax gets a NAN if
-      // one or more item are large enough. a math transform as below is
-      // leveraged to get a stable softmax: e^xi/(e^x1 + ...e^xn) = e^(xi -
-      // max) / (e^(x1 - max) + ... + e^(xn - max))
-      float max = -std::numeric_limits<float>::infinity();
-      for (int i = 0; i < D; i++) {
-        if (max < x[i])
-          max = x[i];
-      }
-      for (int i = 0; i < D; i++) {
-        y[i] = expf(x[i] - max);
-      }
-
-      double sum = 0.0;
-
-      for (int i = 0; i < D; i++) {
-        sum += x[i];
-      }
-
-      if (sum == 0) {
-        for (int i = 0; i < D; i++) {
-          y[i] = 1.0f / (float)D;
-        }
-      } else {
-        for (int i = 0; i < D; i++) {
-          y[i] = x[i] / (float)sum;
-        }
-      }
-    }
-  });
-}
-
-template <>
-inline void ComputeAttentionSoftmaxInplace(float* score, int N, int D, ThreadPool* tp) {
+inline void ComputeAttentionSoftmaxInplace(T* score, int N, int D, ThreadPool* tp) {
   MlasComputeSoftmax(score, score, N, D, false, false, tp);
 }
 
 template <typename T>
-void ComputeAttentionSoftcapInplace(T* scores, int sequence_length, float softcap) {
+void ComputeAttentionSoftcapInplace(T* scores, int sequence_length, T softcap);
+
+template <>
+void ComputeAttentionSoftcapInplace<float>(float* scores, int sequence_length, float softcap) {
   for (int i = 0; i < sequence_length; i++) {
     scores[i] = scores[i] / softcap;
     scores[i] = std::tanh(scores[i]);
@@ -111,7 +38,10 @@ void ComputeAttentionSoftcapInplace(T* scores, int sequence_length, float softca
   }
 }
 
-template void ComputeAttentionSoftcapInplace<float>(float* scores, int sequence_length, float softcap);
+template <>
+void ComputeAttentionSoftcapInplace<MLFloat16>(MLFloat16* scores, int sequence_length, MLFloat16 softcap) {
+  MlasComputeSoftcap(scores, scores, sequence_length, softcap);
+}
 
 template <typename T>
 void PrepareMask(const int32_t* mask_index,