--- a/onnxruntime/core/mlas/lib/rotary_embedding_kernel_neon_fp16.cpp
+++ b/onnxruntime/core/mlas/lib/rotary_embedding_kernel_neon_fp16.cpp
@@ -226,6 +226,7 @@ RopeKernel_Fp16_Impl<true>(
 
 }  // namespace
 
+// TODO(fajin): intra-loop parallelism
 void
 RopeKernel_Fp16(
     const MLAS_FP16* input,