--- a/onnxruntime/core/mlas/lib/softmax_kernel_neon_fp16.cpp
+++ b/onnxruntime/core/mlas/lib/softmax_kernel_neon_fp16.cpp
@@ -20,6 +20,7 @@ Abstract:
 #include "softmax.h"
 #include "softmax_kernel_neon.h"
 
+// TODO(fajin): intra-loop parallelism
 namespace softmax_neon {
 
 template <typename T>